Version 17
I've written my own clex.c to find files, binaries, usernames, and have
added it to the shell. `Nameopt' works, but I'm not sure how to indicate
the first word on the line. `Complete' is probably buggy, I haven't
installed it yet. Changed yankprev to return the 1st position, and so
bug a) is fixed. I am using a `dummy' path at the moment, and should
use the `real' one.

Fixed complete to work, which it appears to do. Also I am now using the
real path, and that works as well. I even made sure I was free()ing
all the malloc'd stuff!! Ta da!