Version 19
Well, parse now parses the carray, and meta now puts things in there, and
qsorts after a *?[], so it's all rather nice. I think gettoken needs a bit
of a rewrite as I just hacked it a lot, and it dies when using ""s. I don't
know if I'll have to insert into the carray when expanding all the meta's,
or just append as I'm doing now. Oh well, we'll soon see.
Altered clex to see ~ as a directory when there is a / in it, simple. I
should write a meta_1 to expand ~ as well I guess.