Version 25
Meta_1 now breaks the line into wordlets, where a wordlet may end in a space,
or be enclosed in "" or '', and later ``. To ensure spaces are preserved, the
carray mode is now a bitfield, with the bit C_SPACE indicating if there's a
space. So       echo fred cat'jim'dog is the list:

        echo    C_SPACE
        fred    C_SPACE
        cat
        'jim'
        dog

I had to change most of meta to use C_SPACE, and expline to expand when
it saw C_SPACE, but now I can get meta_2 & meta_3 to properly parse the
carray. It seems to work quite well, surprisingly. Also sat down and
drew the routine connectivity on a file-by-file basis, found a few routines
I don't use yet. Changed main() so that lines starting with # are ignored
totally, much better.

Clam appears to expand `` before *, so I must do that too.

ToDo: When replacing a word by its meta expansion, make sure the
last part of the expansion has the old C_SPACE bit, and all the other
words in the expansion have C_SPACE on.

Other ideas: (I should start a separate list)

        Make list of things in sh, csh, tcsh & ksh worth adding
        Shell debugger - extra builtins like step & debug
        Add aliases & the prompt variable
        Ifdefing - CLE, ALIAS, SCRIPT HIST
        Add a dumb CLE for file reading
        Proper binding: bind string string
        Tidy up code, esp. comlined & job
        Remove static limits on vars & add dimensionality
        Add other sh $ things
        Change header so only specify machine type on compile line
        $precmd, maybe?

Just cut out and added Callum's prprompt() to main.c, and it works!
I just can't do prompt='string', because '' fails for some reason.
Will freeze this version, as I will start on aliases in the next.